# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Profo 房产数据中心",
    "description": "轻量级、本地化、高性能的房产数据仓库系统",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/renovation": {
      "put": {
        "tags": [
          "projects",
          "projects",
          "renovation"
        ],
        "summary": "Update Renovation Stage",
        "description": "更新改造阶段 (Sync)",
        "operationId": "update_renovation_stage_api_v1_projects__project_id__renovation_put",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "项目ID",
              "title": "Project Id"
            },
            "description": "项目ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenovationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/renovation/photos": {
      "post": {
        "tags": [
          "projects",
          "projects",
          "renovation"
        ],
        "summary": "Upload Renovation Photo",
        "description": "上传改造阶段照片 (Sync)",
        "operationId": "upload_renovation_photo_api_v1_projects__project_id__renovation_photos_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "项目ID",
              "title": "Project Id"
            },
            "description": "项目ID"
          },
          {
            "name": "stage",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "改造阶段",
              "title": "Stage"
            },
            "description": "改造阶段"
          },
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "图片URL",
              "title": "Url"
            },
            "description": "图片URL"
          },
          {
            "name": "filename",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "文件名",
              "title": "Filename"
            },
            "description": "文件名"
          },
          {
            "name": "description",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "描述",
              "title": "Description"
            },
            "description": "描述"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenovationPhotoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "projects",
          "projects",
          "renovation"
        ],
        "summary": "Get Renovation Photos",
        "description": "获取改造阶段照片 (Sync)",
        "operationId": "get_renovation_photos_api_v1_projects__project_id__renovation_photos_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "项目ID",
              "title": "Project Id"
            },
            "description": "项目ID"
          },
          {
            "name": "stage",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "改造阶段筛选",
              "title": "Stage"
            },
            "description": "改造阶段筛选"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  },
                  "title": "Response Get Renovation Photos Api V1 Projects  Project Id  Renovation Photos Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/renovation/photos/{photo_id}": {
      "delete": {
        "tags": [
          "projects",
          "projects",
          "renovation"
        ],
        "summary": "Delete Renovation Photo",
        "description": "删除改造阶段照片 (Sync)",
        "operationId": "delete_renovation_photo_api_v1_projects__project_id__renovation_photos__photo_id__delete",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "项目ID",
              "title": "Project Id"
            },
            "description": "项目ID"
          },
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "照片ID",
              "title": "Photo Id"
            },
            "description": "照片ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/renovation/contract": {
      "get": {
        "tags": [
          "projects",
          "projects",
          "renovation"
        ],
        "summary": "Get Renovation Contract",
        "description": "获取装修合同信息 (Sync)",
        "operationId": "get_renovation_contract_api_v1_projects__project_id__renovation_contract_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "项目ID",
              "title": "Project Id"
            },
            "description": "项目ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenovationContractResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "projects",
          "projects",
          "renovation"
        ],
        "summary": "Update Renovation Contract",
        "description": "更新装修合同信息 (Sync)",
        "operationId": "update_renovation_contract_api_v1_projects__project_id__renovation_contract_put",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "项目ID",
              "title": "Project Id"
            },
            "description": "项目ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenovationContractUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenovationContractResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "项目ID"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "项目名称"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "项目状态"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "community_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Community Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area"
          },
          "layout": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Layout"
          },
          "orientation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orientation"
          },
          "is_deleted": {
            "type": "boolean",
            "title": "Is Deleted",
            "default": false
          },
          "renovation_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Renovation Stage"
          },
          "contract_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract No",
            "description": "合同编号"
          },
          "signing_price": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signing Price",
            "description": "签约价格(万)"
          },
          "signing_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signing Date"
          },
          "signing_period": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signing Period"
          },
          "extension_period": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extension Period"
          },
          "extension_rent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extension Rent"
          },
          "cost_assumption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Assumption"
          },
          "planned_handover_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Handover Date"
          },
          "other_agreements": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Other Agreements"
          },
          "contract_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Status"
          },
          "owner_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Name"
          },
          "owner_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Phone"
          },
          "owner_id_card": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id Card"
          },
          "owner_info": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Info"
          },
          "list_price": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "List Price",
            "description": "挂牌价(万)"
          },
          "listing_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listing Date"
          },
          "sold_price": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sold Price"
          },
          "sold_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sold Date"
          },
          "transaction_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Status"
          },
          "total_income": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Income",
            "default": "0"
          },
          "total_expense": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Expense",
            "default": "0"
          },
          "net_cash_flow": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Cash Flow",
            "default": "0"
          },
          "roi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roi",
            "default": 0.0
          },
          "signing_materials": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SigningMaterial"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signing Materials",
            "description": "签约材料列表"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectResponse",
        "description": "项目完整响应模型 - 适配新的规范化表结构\n业务日期字段使用字符串类型 (YYYY-MM-DD) 避免时区问题"
      },
...
      "RenovationPhotoResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "stage",
          "url",
          "created_at"
        ],
        "title": "RenovationPhotoResponse",
        "description": "照片响应"
      },
      "RenovationUpdate": {
        "properties": {
          "renovation_company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Renovation Company"
          },
          "contract_start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Start Date"
          },
          "contract_end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract End Date"
          },
          "actual_start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Start Date"
          },
          "actual_end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual End Date"
          },
          "hard_contract_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hard Contract Amount"
          },
          "payment_node_1": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Node 1"
          },
          "payment_ratio_1": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Ratio 1"
          },
          "payment_node_2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Node 2"
          },
          "payment_ratio_2": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Ratio 2"
          },
          "payment_node_3": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Node 3"
          },
          "payment_ratio_3": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Ratio 3"
          },
          "payment_node_4": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Node 4"
          },
          "payment_ratio_4": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Ratio 4"
          },
          "soft_budget": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Soft Budget"
          },
          "soft_actual_cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Soft Actual Cost"
          },
          "soft_detail_attachment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Soft Detail Attachment"
          },
          "design_fee": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Design Fee"
          },
          "demolition_fee": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Demolition Fee"
          },
          "garbage_fee": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Garbage Fee"
          },
          "other_extra_fee": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Other Extra Fee"
          },
          "other_fee_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Other Fee Reason"
          }
        },
        "type": "object",
        "title": "RenovationUpdate",
        "description": "更新装修记录请求"
      },
...
      "SigningMaterial": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename",
            "description": "文件名"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "文件URL"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "description": "附件分类"
          },
          "fileType": {
            "type": "string",
            "title": "Filetype",
            "description": "文件类型"
          },
          "size": {
            "type": "integer",
            "title": "Size",
            "description": "文件大小(字节)",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "filename",
          "url",
          "category",
          "fileType"
        ],
        "title": "SigningMaterial",
        "description": "签约材料附件"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        }
      }
    }
  }
}